[backend] refactor: 팀의 멤버 이름 조회 로직 계층 분리 및 구조 개선 (Controller → Service → Repository → Mapper)

+++ backend/src/main/java/com/example/demo/member/MemberMapper.java
@@ -22,7 +22,7 @@ public interface MemberMapper {
 
     void deleteMember(long memberId);    // 멤버 삭제
 
-    List<TeamMembersNameResponse> requestTeamMembersName(int teamId);  // 팀의 모든 멤버 이름 요청
+    List<TeamMembersNameResponse> requestTeamMembersName(long teamId);  // 팀의 모든 멤버 이름 요청
 
     void requestInviteMember(MemberInviteInTeamRequestDTO member);   // 팀에 멤버를 초대
